+++ Sources/SwifterSwiftUIKitCore/UICollectionViewExtensions.swift
@@ -143,7 +143,17 @@ public extension UICollectionView {
         if let bundleName = bundleClass {
             bundle = Bundle(for: bundleName)
         }
+        register(UINib(nibName: identifier, bundle: bundle), forCellWithReuseIdentifier: identifier)
+    }
 
+    /// SwifterSwift: Register UICollectionViewCell with .xib file using only its corresponding class.
+    ///               Assumes that the .xib filename and cell class has the same name.
+    ///
+    /// - Parameters:
+    ///   - name: UICollectionViewCell type.
+    ///   - bundle: Bundle where the compiled nib exists.
+    func register(nibWithCellClass name: (some UICollectionViewCell).Type, bundle: Bundle?) {
+        let identifier = String(describing: name)
         register(UINib(nibName: identifier, bundle: bundle), forCellWithReuseIdentifier: identifier)
     }