+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The changelog for **SwifterSwift**. Also see the [releases](https://github.com/S
   - Added `numberOfDaysInYear(for:)` method to get the total number of days in a date's year. [#1253](https://github.com/SwifterSwift/SwifterSwift/pull/1253) by [William-Laverty](https://github.com/William-Laverty)
 - **Xcode**
   - Added support for Xcode 26. [#1240](https://github.com/SwifterSwift/SwifterSwift/pull/1240) by [guykogus](https://github.com/guykogus)
+- **UIKit**
+  - Added `UICollectionView.register(nibWithCellClass:bundle:)` to register nib-based cells from an explicit bundle. [#1187](https://github.com/SwifterSwift/SwifterSwift/pull/1187) by [guykogus](https://github.com/guykogus)
+  - Added `UITableView.register(nibWithCellClass:bundle:)` to register nib-based cells from an explicit bundle. [#1187](https://github.com/SwifterSwift/SwifterSwift/pull/1187) by [guykogus](https://github.com/guykogus)
 ### Changed
   - `Date.string(withFormat:)` now support locale and timeZone parameter. [#1250](https://github.com/SwifterSwift/SwifterSwift/pull/1250) by [weihas](https://github.com/weihas)
 ### Fixed